requirements.txt
ConfigArgParse
einops
Flask
huggingface-hub
imageio
loguru
matplotlib
natsort
nltk
opencv-python
pandas
Pillow
protobuf
pybullet
pycocotools
requests
scikit-image
scikit-learn
scipy
six
spacy
tdqm
tensorboard
timm
tokenizers
torch
torchvision
transformers